Move shared components to themedComponents
